fixed crash in soap_device.c, I guess I screwed up when adding custom header support.
fixed content directory service subscriptions after check with 
DeviceValidator from Intel UPnP Tools

server.cc: we were overwriting the error code

+++ mediatomb/tombupnp/upnp/src/genlib/net/http/httpparser.c
@@ -870,7 +870,7 @@ match_int( INOUT scanner_t * scanner,
     memptr token;
     token_type_t tok_type;
     parse_status_t status;
-    int num;
+    long num;
     char *end_ptr;
     size_t save_pos;
